Match the ccache size to jenkins cache (#97)
ccache has a default size of 5GB and will keep the cache below that  value. If the Jenkins cache is smaller (e.g. 4GB) then the cache will be discarded any time it goes above 4GB. Instead we can decrease the ccache size so that the cache should always fit.